Refactor package repository defintition in dedicated file
