+++ imagecraft/models/__init__.py
@@ -15,6 +15,7 @@
 # with this program.  If not, see <http://www.gnu.org/licenses/>.
 """Data models for Imagecraft."""
 
-from imagecraft.models.project import Project, Platform, PackageRepository
+from imagecraft.models.project import Project, Platform
+from imagecraft.models.package_repository import PackageRepository
 
 __all__ = ["Project", "Platform", "PackageRepository"]
