+++ imagecraft/models/project.py
@@ -19,8 +19,6 @@
 This module defines a imagecraft.yaml file, exportable to a JSON schema.
 """
 
-import enum
-import re
 from collections.abc import Mapping
 from typing import TYPE_CHECKING, Any
 
@@ -29,28 +27,9 @@ from craft_application.errors import CraftValidationError
 from craft_application.models import BuildInfo
 from craft_application.models import Project as BaseProject
 from craft_application.util.error_formatting import format_pydantic_errors
-from craft_archives.repo import errors  # type: ignore[import-untyped]
-from craft_archives.repo.package_repository import (  # type: ignore[import-untyped]
-    KeyIdStr,
-)
-
-# pyright: reportMissingTypeStubs=false
-from craft_archives.repo.package_repository import (  # type: ignore[import-untyped]
-    PackageRepository as BasePackageRepository,
-)
-from craft_archives.repo.package_repository import (
-    PackageRepositoryApt as BasePackageRepositoryApt,
-)
-from craft_archives.repo.package_repository import (
-    PackageRepositoryAptPPA as BasePackageRepositoryAptPPA,
-)
-from craft_cli import CraftError
 from craft_providers import bases
 from pydantic import (
-    AnyUrl,
     BaseModel,
-    ConstrainedStr,
-    FileUrl,
     ValidationError,
     conlist,
     root_validator,  # pyright: ignore[reportUnknownVariableType]
@@ -59,6 +38,12 @@ from pydantic import (
 from pydantic_yaml import YamlModelMixin
 
 from imagecraft.architectures import SUPPORTED_ARCHS
+from imagecraft.models.errors import ProjectValidationError
+from imagecraft.models.package_repository import (
+    PackageRepository,
+    PackageRepositoryApt,
+    PackageRepositoryPPA,
+)
 
 # A workaround for mypy false positives
 # see https://github.com/samuelcolvin/pydantic/issues/975#issuecomment-551147305
@@ -73,29 +58,6 @@ file_name = "imagecraft.yaml"
 def _alias_generator(s: str) -> str:
     return s.replace("_", "-")
 
-
-class AuthStr(ConstrainedStr):
-    """A constrained string for a auth string."""
-
-    regex = re.compile(r".*:.*")
-
-class UsedForEnum(enum.Enum):
-    """Enum values that represent how/when the package repository can be used for."""
-
-    BUILD = "build"
-    RUN = "run"
-    ALWAYS = "always"
-
-class PocketEnum(enum.Enum):
-    """Enum values that represent possible pocket values."""
-
-    RELEASE = "release"
-    UPDATES = "updates"
-    PROPOSED = "proposed"
-    SECURITY = "security"
-
-class ProjectValidationError(CraftError):
-    """Error validating imagecraft.yaml."""
 
 class ProjectModel(YamlModelMixin, BaseProject):
     """Base model for the imagecraft project class."""
@@ -171,82 +133,6 @@ class Platform(ElementModel):
             )
 
         return values
-
-class PackageRepository(BasePackageRepository): # type:ignore[misc]
-    """Imagecraft package repository definition."""
-
-    @classmethod
-    def unmarshal(cls, data: Mapping[str, Any]) -> "PackageRepositoryPPA | PackageRepositoryApt":
-        """Create a package repository object from the given data."""
-        if not isinstance(data, dict):  # pyright: ignore[reportUnnecessaryIsInstance]
-            raise errors.PackageRepositoryValidationError(
-                url=str(data),
-                brief="invalid object.",
-                details="Package repository must be a valid dictionary object.",
-                resolution=(
-                    "Verify repository configuration and ensure that the "
-                    "correct syntax is used."
-                ),
-            )
-
-        if "ppa" in data:
-            return PackageRepositoryPPA.unmarshal(data)
-
-        return PackageRepositoryApt.unmarshal(data)
-
-class PackageRepositoryPPA(BasePackageRepositoryAptPPA): # type:ignore[misc]
-    """Imagecraft PPA repository definition."""
-
-    auth: AuthStr | None
-    key_id: KeyIdStr | None = pydantic.Field(alias="key-id") # here until added to craft-archives
-    used_for: UsedForEnum = UsedForEnum.ALWAYS
-
-    @classmethod
-    def unmarshal(cls, data: Mapping[str, Any]) -> "PackageRepositoryPPA":
-        """Create and populate a new ``PackageRepository`` object from dictionary data.
-
-        The unmarshal method validates entries in the input dictionary, populating
-        the corresponding fields in the data object.
-
-        :param data: The dictionary data to unmarshal.
-
-        :return: The newly created object.
-
-        :raise TypeError: If data is not a dictionary.
-        """
-        if not isinstance(data, dict): # pyright: ignore[reportUnnecessaryIsInstance]
-            raise TypeError("Package repository PPA data is not a dictionary")
-
-        return cls(**data)
-
-class PackageRepositoryApt(BasePackageRepositoryApt): # type:ignore[misc]
-    """Imagecraft APT package repository definition."""
-
-    url: AnyUrl | FileUrl | None # pyright: ignore[reportIncompatibleVariableOverride]
-    key_id: KeyIdStr | None = pydantic.Field(alias="key-id") # pyright: ignore[reportIncompatibleVariableOverride]
-    used_for: UsedForEnum = UsedForEnum.ALWAYS
-
-    pocket: PocketEnum | None = PocketEnum.RELEASE
-    flavor: str | None
-
-    @classmethod
-    def unmarshal(cls, data: Mapping[str, Any]) -> "PackageRepositoryApt":
-        """Create and populate a new ``PackageRepositoryApt`` object from dictionary data.
-
-        The unmarshal method validates entries in the input dictionary, populating
-        the corresponding fields in the data object.
-
-        :param data: The dictionary data to unmarshal.
-
-        :return: The newly created object.
-
-        :raise TypeError: If data is not a dictionary.
-        """
-        if not isinstance(data, dict): # pyright: ignore[reportUnnecessaryIsInstance]
-            raise TypeError("Package repository data is not a dictionary")
-
-        return cls(**data)
-
 
 class Project(ProjectModel):
     """Definition of imagecraft.yaml configuration."""
